// client-v2/src/interfaces/graph.ts
export interface GraphDatasetsArgs {
    data: any[];
    borderColor: string;
    backgroundColor: string| any;
    label?: boolean;
    fill?: boolean;
    borderWidth?: number;
    tension?: number;
    pointRadius?: number;
    borderRadius?: number;
}

/**
 * Interface which represents the the result of graph datasets
 */
export interface GraphDatasetsResult {
    label: boolean;
    data: { x: number; y: number }[];
    fill: boolean;
    borderColor: string;
    borderWidth: number;
    backgroundColor: string;
    tension: number;
    pointRadius: number;
    borderRadius?: number;
}

/**
 * Interface for configuring the scales in a graph
 */
export interface GraphScalesDataArgs {
    titleText?: string;
    titleColor?: string;
    ticksColor?: string;
    gridDisplay?: boolean;
    titleDisplay?: boolean;
    maxRotation?: number;
    maxTicksLimit?: number;
    borderColor?: string;
    position?: string;
}

/**
 * Interface which represents the result of graph scales
 */
export interface GraphScalesResult {
    position: string;
    grid: {
      display: boolean;
      borderColor?: string;
    };
    title: {
      display: boolean;
      text?: string;
      color?: string;
    };
    ticks: {
      maxRotation: number;
      maxTicksLimit: number;
      color?: string;
    };
}

/**
 * Interface for configuring the options of a graph
 */
export interface GraphOptionsArgs {
    responsive?: boolean;
    maintainAspectRatio?: boolean;
    cutoutPercentage?: number;
    legendDisplay?: boolean;
}

/**
 * Interface which represents the the result of graph options
 */
export interface GraphOptionsResult {
    responsive: boolean;
    maintainAspectRatio: boolean;
    cutoutPercentage: number;
    plugins?: PluginsOptions;
    hover: HoverOptions;
}

/**
 * Interface for configuring the options of a graph Tooltip
 */
export interface TooltipGraphOptionsArgs {
    hoverBackgroundColor?: string;
    boxPadding?: number;
    padding?: number;
    size?: number;
    mode?: string;
    intersect?: boolean;
}

/**
 * Interface which represents the the result of graph Tooltip options
 */
export interface TooltipGraphOptionsResult {
    hoverBackgroundColor: string;
    hover: {
      mode: string;
      intersect: boolean;
    };
    plugins: {
      tooltip: {
        bodyFont: {
          size: number;
        };
        titleFont: {
          size: number;
        };
        boxPadding: number;
        padding: number;
      };
    };
}

/**
 * Interface for configuring the oplugin options for graph options
 */
export interface PluginsOptions {
    legend?: LegendOptions;
}

/**
 * Interface for configuring the legend options for graph options
 */
export interface LegendOptions {
    display?: boolean;
}

/**
 * Interface for configuring the hover options for graph options.
 */
export interface HoverOptions {
    mode: null;
}